Only importing matplotlib when needed
If DISPLAY wasn't set, I got an error

  QXcbConnection: Could not connect to display

in ipython, after which the process ended. This was apparently due
to the matplotlib.pyplot import. To avoid a script from crashing when
some functions of plotutil are needed, but matplotlib is not used,
now the module is only imported when needed.