Bugfix: bad sinh/cosh functions for windows
This was a remnant from way way back when I wanted the old
grale libs to compile on windows; it was never really used in
windows however, and the buggy nature of these function definitions
never got noticed.

Now it's just all STL, should be supported on all compilers used.