Bugfix: base range of uniform number is wrong, should be -1..1
This goes way back to 2008, when the code was committed. The existing
version caused a random number between 0 and 1 to be generated,
resulting in changes (in the small mutation phase) that can only
lead to a larger mass sheet.

At this point I'm not certain if that was done for a reason or by
mistake: on the one hand it seems like a copy paste of another piece
of code, but with a small change. That could imply it was deliberate.
It is also possible that I only copied part of the code, and typed
an incomplete first line.

It does seem that this should be changed, which is what I've done. The
effect should certainly be investigated more closely.