Spaces/tab issue, vscode detected it wrong
Apparently vscode guesses the use of spaces/tabs, it is not linked
to a specific file type. For new files the preference is spaces,
and for existing files it is sometimes detected wrong (due to a
license header?). In this, and the following commits, indentation
has been replaced to tabs again.