Problem: using io.readBytes can be too slow and trigger timeout
Especially in timedio, things are read bytes per byte to be able to
check the timeouts, which is way too slow for more complex lenses
(or DeflectionGridLens for example) and still allows the timeout
to be triggered.

Added a readBytesUntimed function for these results, which does not
do any timing checks.

A better way to handle this would be to re-implement this using C/C++
so that timeouts can still be used but more than one byte at a time is
read.