Copying cppqt lib file to egg dir if necessary
If seems that in a recent installation setuptools overrides
distutil and creates an egg directory for installation. This causes
the 'make install' of cppqt to end up in the wrong directory. The
setup.py script now checks if the installation dir is of the egg type,
and copies the cppqt lib to the appropriate directory