Quick test code to check if a difference was caused by difference in prior
A parametric inversion test with MCMC sampling seemed to yield different
results for the marginalized distributions of the coreradius variable,
for a PIEMDLens vs an LTPIEMLens. Apart from the coreradius variable,
the centraldensity or velocitydispersion was also allowed to change.

In these optimisations, uniform prior distributions on the parameters
were assumed, but a uniform prior on (centraldensity,coreradius) is
not a uniform density on (velocitydispersion,coreradius) and vice versa.
To check that this was indeed the cause of the difference in
marginalized distributions, the prior to add to a LTPIEMDLens model
was calculated and added, and then indeed the same results are obtained
as when using a PIEMLens.