Don't use gaussian prior for derived cname variables
In the current approach of calculating the priors this is not important,
they are only calculated for the variables that do change. I'm thinking
of changing the calculation to also allow for other opencl code to run,
and if that were used to calculate the priors as well, it could become
an issue.